Удалит моковые данные.
