+++ src/components/sort-type-list.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import { useAppDispatch, useAppSelector } from '../hooks';
-import { SortTypeName } from '../mocks/mock-const';
 import SortType from './sort-type';
 import { changeSortType, sortCityOffers } from '../store/actions';
