+++ src/const.ts
@@ -10,7 +10,14 @@ const UrlMarker = {
   Selected: 'img/pin-active.svg'
 } as const;
 
+const SortTypeName = {
+  Popular: 'Popular',
+  PriceLowToHigh: 'Price: low to high',
+  PriceHighToLow: 'Price: high to low',
+  TopRatedFirst: 'Top rated first'
+};
+
 const RATING_TO_BAR_WIDTH_RATIO = 20;
 
-export { Path, UrlMarker, RATING_TO_BAR_WIDTH_RATIO };
+export { Path, UrlMarker, RATING_TO_BAR_WIDTH_RATIO, SortTypeName };
