--- src/mocks/cities.ts
@@ -1,52 +1,0 @@
-import { CityName } from './mock-const';
-
-export const CityInfo = [
-  {
-    title: CityName.Paris,
-    location: {
-      lat: 52.35514938496378,
-      lng: 4.673877537499948,
-      zoom: 7
-    },
-  },
-  {
-    title: CityName.Amsterdam,
-    location: {
-      lat: 52.35514938496378,
-      lng: 4.673877537499948,
-      zoom: 9
-    },
-  },
-  {
-    title: CityName.Brussels,
-    location: {
-      lat: 52.35514938496378,
-      lng: 4.673877537499948,
-      zoom: 6
-    },
-  },
-  {
-    title: CityName.Cologne,
-    location: {
-      lat: 52.35514938496378,
-      lng: 4.673877537499948,
-      zoom: 5
-    },
-  },
-  {
-    title: CityName.Dusseldorf,
-    location: {
-      lat: 52.35514938496378,
-      lng: 4.673877537499948,
-      zoom: 10
-    },
-  },
-  {
-    title: CityName.Hamburg,
-    location: {
-      lat: 52.35514938496378,
-      lng: 4.673877537499948,
-      zoom: 8
-    },
-  }
-];
