--- src/mocks/comments.ts
@@ -1,117 +1,0 @@
-import { UserComment } from '../types/comment';
-
-const UserComments: UserComment[] = [
-  {
-    id: '1',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Антон',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Lorem ipsum dolor sit amet.',
-    rating: 5,
-  },
-  {
-    id: '2',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Анна',
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    comment: 'Nam quam nunc, blandit ve.',
-    rating: 4,
-  },
-  {
-    id: '3',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Борис',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Sed consequat, leo eget bibendum sodales, augue velit cursus nunc.',
-    rating: 3.9,
-  },
-  {
-    id: '4',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Владимир',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Maecenas nec odio et ante tincidunt tempus. Donec vitae.',
-    rating: 4.9,
-  },
-  {
-    id: '5',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Ольга',
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    comment: 'Maecenas tempus, tellus eget condimentum rhoncus, sem quam semper libero.',
-    rating: 4.2,
-  },
-  {
-    id: '6',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Геннадий',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Quisque rutrum. Aenean imperdiet.',
-    rating: 3.8,
-  },
-  {
-    id: '7',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Александр',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Etiam rhoncus.',
-    rating: 4,
-  },
-  {
-    id: '8',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Екатерина',
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    comment: 'Curabitur ullamcorper.',
-    rating: 3.5,
-  },
-  {
-    id: '9',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Алина',
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    comment: 'Aenean massa.',
-    rating: 4.5,
-  },
-  {
-    id: '10',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Сергей',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Donec pede justo, fringilla vel.',
-    rating: 4.3,
-  },
-  {
-    id: '11',
-    date: '2019-05-08T14:13:56.569Z',
-    user: {
-      name: 'Афанасий',
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    comment: 'Integer tincidunt. Cras dapibus.',
-    rating: 5,
-  },
-];
-
-export { UserComments };
-
