--- src/mocks/mock-const.ts
@@ -1,24 +1,0 @@
-const CityName = {
-  Paris: 'Paris',
-  Cologne: 'Cologne',
-  Brussels: 'Brussels',
-  Amsterdam: 'Amsterdam',
-  Hamburg: 'Hamburg',
-  Dusseldorf: 'Dusseldorf'
-};
-
-const SortTypeName = {
-  Popular: 'Popular',
-  PriceLowToHigh: 'Price: low to high',
-  PriceHighToLow: 'Price: high to low',
-  TopRatedFirst: 'Top rated first'
-};
-
-const enum HousingType {
-  Apartment = 'apartment',
-  Hotel = 'hotel',
-  House = 'house',
-  Room = 'room'
-}
-
-export { CityName, SortTypeName, HousingType };
